+++ backend/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+Database schema version: 43
+
+## Added
+
+- Added `token_module_pause_unpause_status` column to `plt_tokens` query to get the current status of the token module (paused or unpaused).
+
 ## [2.0.20] - 2025-09-10
 
 Database schema version: 41